Add files via upload
The main objective of the project is to produce an adequate prediction model that helps predict credit defaults using explanatory variables identified in the dataset. To achieve this, the project will compare various classification methods to identify and forecast loan defaulters. The methods we used are logistic regression, LDA, QDA, Naïve Bayes, and KNN for the original data as well as Regularization methods, such as LASSO, Ridge and Elastic Net. In addition, logistic regression, LDA, QDA, Naïve Bayes, and KNN were performed on the principal components of the explanatory variables. These models were trained on a training dataset, which is 70% of the entire dataset, and tested on the remaining holdout dataset for accuracy and sensitivity. Finally, a champion model, with the smallest test error, is selected as the model of choice.